Feat: made main settings window
