+++ lib/pages/general/menu_pages/settings/typeSettings/type_tile.dart
